[docs] remove didReceiveAttrs lifecycle hook
We want our components to more closely match Glimmer Components which
will not have the `didReceiveAttrs` lifecycle hook.

We also set `tagName` to an empty string which makes this component
not have a wrapping element. This is also the default for Glimmer
Components so we'll want this example to match that.